+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/di/KoinModule.kt
@@ -2,9 +2,8 @@ package com.altmann.choresmanager.di
 
 import com.altmann.choresmanager.database.UserDao
 import com.altmann.choresmanager.local.LocalDatabase
-import com.altmann.choresmanager.ui.screens.addchorepopup.ChorePopupViewModel
+import com.altmann.choresmanager.ui.screens.chore.addchorepopup.ChorePopupViewModel
 import com.altmann.choresmanager.ui.screens.calendar.CalendarViewModel
-import com.altmann.choresmanager.ui.screens.home.HomeScreen
 import com.altmann.choresmanager.ui.screens.sidebar.SideBarViewModel
 import com.altmann.choresmanager.viewmodels.SharedChoreViewModel
 import org.koin.core.KoinApplication