+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/calendar/CalendarUiState.kt
@@ -7,7 +7,8 @@ import kotlinx.datetime.LocalDate
 data class CalendarUiState(
     val anchor: LocalDate = LocalDate(1970,1,1),
     val selectedDate : LocalDate = CalendarHelper.today(),
-    val expanded : Boolean = false,
+    val expandedAddChore : Boolean = false,
+    val expandedChore : Boolean = false,
     val choresByDate : Map<LocalDate, List<Chore>> = emptyMap(),
     val enabledChores : List<Chore> = emptyList()
 )