+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/chore/addchorepopup/ChorePopupViewModel.kt
@@ -1,6 +1,5 @@
-package com.altmann.choresmanager.ui.screens.addchorepopup
+package com.altmann.choresmanager.ui.screens.chore.addchorepopup
 
-import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow